+++ packages/solid/src/components/avatar/index.ts
@@ -1,4 +1,5 @@
 export * from "./avatar";
+export * from "./avatar.styles";
 export * from "./avatar-badge";
 export * from "./avatar-excess";
 export * from "./avatar-group";