+++ packages/solid/src/components/circular-progress/index.ts
@@ -1,3 +1,4 @@
 export * from "./circular-progress";
+export * from "./circular-progress.styles";
 export * from "./circular-progress-indicator";
 export * from "./circular-progress-label";