+++ packages/solid/src/components/collapse/index.ts
@@ -1,1 +1,2 @@
 export * from "./collapse";
+export * from "./collapse.styles";