+++ packages/solid/src/components/icon/index.ts
@@ -1,2 +1,3 @@
 export * from "./create-icon";
 export * from "./icon";
+export * from "./icon.styles";