+++ packages/solid/src/components/modal/index.ts
@@ -1,4 +1,5 @@
 export * from "./modal";
+export * from "./modal.styles";
 export * from "./modal-body";
 export * from "./modal-close-button";
 export * from "./modal-content";