+++ packages/solid/src/components/skeleton/index.ts
@@ -1,3 +1,4 @@
 export * from "./skeleton";
+export * from "./skeleton.styles";
 export * from "./skeleton-circle";
 export * from "./skeleton-text";