+++ packages/solid/src/components/tag/index.ts
@@ -1,4 +1,5 @@
 export * from "./tag";
+export * from "./tag.styles";
 export * from "./tag-close-button";
 export * from "./tag-icon";
 export * from "./tag-label";